pyproject.toml
[project]
name = "email-labeling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.23.0",
]
